requirements.txt
python-pptx==0.6.18
Pillow==10.2.0
moviepy==1.0.3
